python/requirements.txt
async-timeout==4.0.2
importlib-metadata==6.6.0
pkg-resources==0.0.0
pyserial==3.5
python-dotenv==0.21.1
pytz==2023.3
redis==4.5.5
typing-extensions==4.5.0
zipp==3.15.0
